chore: update Reward page
