+++ apps/frontend/src/app/3_organisms/RewardHistory/types.ts
@@ -1,0 +1,9 @@
+export enum RewardHistoryType {
+  stabilityPoolRewards = 'stabilityPoolRewards',
+  stabilityPoolSubsidies = 'stabilityPoolSubsidies',
+}
+
+export type RewardHistoryProps = {
+  selectedHistoryType: RewardHistoryType;
+  onChangeRewardHistory: (value: RewardHistoryType) => void;
+};
